refactor(test): fold btc-bs58-override into btc-pair
Both files target the `pair_ledger_btc` flow. The bs58/base-x subtree
test reaches into `@ledgerhq/hw-app-btc` via runtime `createRequire`
to assert that bs58check's decode returns a real Node Buffer (issue
#181 + sibling). It does NOT mock `@ledgerhq/hw-app-btc` or `bs58`,
so the file-level mocks already in btc-pair.test.ts (btc-usb-loader,
btc indexer) don't affect it — folding is safe.

The marinade-related single-test files (anchor-bn-import-shape,
solana-marinade-sdk-resolution) intentionally stay standalone:
solana-marinade.test.ts mocks `@coral-xyz/anchor` and the marinade
SDK at module level (vi.mock is hoisted file-wide), and folding
either would mask the very regression they guard against.

1116 tests / 91 files (was 92).
